Refine dashboard activity report experience
